Manually added test differences
