+++ app/controllers/events_controller.rb
@@ -1,19 +1,14 @@
-
 class EventsController < ApplicationController
   before_action :set_event, only: %i[ show edit update destroy ]
+  
   before_action :authenticate_user!
   before_action :check_officer_privelege
-
+  before_action :generate_qr_attendance
+  before_action :generate_qr_shifts
+  
   # GET /events or /events.json
   def index
-    @TEST = request.base_url
     @events = Event.all
-    puts '\n'
-    puts @TEST
-    puts '\n'
-    if @TEST.nil?
-      puts 'EXISTS'
-    end
   end
 
   def names
@@ -23,13 +18,11 @@ class EventsController < ApplicationController
   # GET /events/1 or /events/1.json
   def show
     @EventShifts = Shift.where(Event_id: params[:id])
-
     @EventAttendances = Attendance.joins(:Shift).where(Shift: {Event_id: params[:id]})
   end
 
   # GET /events/new
   def new
-    @TEST = request.orginal_url
     @event = Event.new
   end
 
@@ -40,7 +33,6 @@ class EventsController < ApplicationController
   # POST /events or /events.json
   def create
     @event = Event.new(event_params)
-
     respond_to do |format|
       if @event.save
         format.html { redirect_to event_url(@event), notice: "Event was successfully created." }
@@ -68,25 +60,12 @@ class EventsController < ApplicationController
   # DELETE /events/1 or /events/1.json
   def destroy
     @event.destroy
-
     respond_to do |format|
       format.html { redirect_to events_url, notice: "Event was successfully destroyed." }
       format.json { head :no_content }
     end
-
-  def CreateQr
-    @qr = RQRCode::QRCode.new(@event.find(params[:id]))
-    puts qr.to_s
-
-    @svg = qr.as_svg(
-      color: "000",
-      shape_rendering: "crispEdges",
-      module_size: 11,
-      standalone: true,
-      use_path: true
-    )
-    end
   end
+
   private
     # Use callbacks to share common setup or constraints between actions.
     def set_event
@@ -96,5 +75,39 @@ class EventsController < ApplicationController
     # Only allow a list of trusted parameters through.
     def event_params
       params.require(:event).permit(:Type, :Date, :Name, :Start, :End)
+    end
+
+    def generate_qr_attendance
+      qrcode_attendance = RQRCode::QRCode.new(request.base_url + "/attendance/new?signup=1" )
+      png = qrcode_attendance.as_png(
+          bit_depth: 1,
+          border_modules: 4,
+          color_mode: ChunkyPNG::COLOR_GRAYSCALE,
+          color: "black",
+          file: nil,
+          fill: "white",
+          module_px_size: 6,
+          resize_exactly_to: false,
+          resize_gte_to: false,
+          size: 120
+      )
+      @png = png.to_s
+    end
+
+    def generate_qr_shifts
+      qrcode_shift = RQRCode::QRCode.new(request.base_url + "/attendances" )
+      png = qrcode_shift.as_png(
+          bit_depth: 1,
+          border_modules: 4,
+          color_mode: ChunkyPNG::COLOR_GRAYSCALE,
+          color: "black",
+          file: nil,
+          fill: "white",
+          module_px_size: 6,
+          resize_exactly_to: false,
+          resize_gte_to: false,
+          size: 120
+      )
+      @png2 = png.to_s   
     end
 end
