+++ app/views/events/index.html.erb
@@ -1,5 +1,14 @@
-<p id="notice"><%= notice %></p>
 <%= stylesheet_link_tag 'events' %>
+<% if flash[:notice] %>
+  <div class="alert alert-success" role="alert" id="hideMeAfter5Seconds" style="position: fixed; right: 1rem;">
+    <%= flash[:notice] %>
+  </div>
+<% end %>
+
+<strong> sign-up </strong>
+<%= image_tag "data:image/png;base64,#{Base64.strict_encode64(@png)}" %>
+<strong> sign-in </strong>
+<%= image_tag "data:image/png;base64,#{Base64.strict_encode64(@png2)}" %>
 
 <div class="d-flex align-items-center gap-3 mb-2">
   <h1 style="margin: 0">Events</h1>
@@ -15,8 +24,6 @@
     <thead>
       <tr>
         <th>Name</th>
-        <th>Shifts</th>
-        <th>Attendance</th>
         <th>Type</th>
         <th>Date</th>
         <th>Start Time</th>
@@ -33,8 +40,6 @@
         <% else %>
           <tr>
             <td><%= link_to event.Name, event, :class => 'name_links', :style => 'color: black'; %></td>
-            <td><%= image_tag(event.qr_code_shift) if event.qr_code_shift.attached? %></td>
-            <td><%= image_tag(event.qr_code_attendance) if event.qr_code_attendance.attached? %></td>
             <%if event.Type == 0 %>
               <td>Club Meeting</td>
             <%else%>
@@ -68,8 +73,6 @@
     <thead>
       <tr>
         <th>Name</th>
-        <th>Shifts</th>
-        <th>Attendance</th>
         <th>Type</th>
         <th>Date</th>
         <th>Start Time</th>
@@ -84,8 +87,6 @@
         <% else %>
           <tr>
             <td><%= link_to event.Name, event, :class => 'name_links', :style => 'color: black'; %></td>
-            <td><%= image_tag(event.qr_code_shift) if event.qr_code_shift.attached? %></td>
-            <td><%= image_tag(event.qr_code_attendance) if event.qr_code_attendance.attached? %></td>
             <%if event.Type == 0 %>
               <td>Club Meeting</td>
             <%else%>
