+++ app/views/events/show.html.erb
@@ -33,29 +33,6 @@
         <%= @event.End.strftime("%I:%M %p") %>
       </p>
     </div>
-    <div class="col-sm">
-      <ul>
-        <% @event.qr_code_attendance do |file| %>
-          <li>
-            <% if file.representable? %>
-              <%= image_tag file.representation(resize_to_limit: [100, 100]) %>
-            <% else %>
-              <%= link_to rails_blob_path(file, disposition: "attachment") do %>
-                <%= image_tag "placeholder.png", alt: "Download file" %>
-              <% end %>
-            <% end %>
-          </li>
-        <% end %>
-      </ul>
-
-      <p><strong>Sign Up:</strong></p>
-      <p><%= image_tag(@event.qr_code_shift)%></p>
-
-    </div>
-    <div class="col-sm">
-      <p><strong>Sign In:</strong></p>
-      <p><%= image_tag(@event.qr_code_attendance) %></p>
-    </div>
   </div>
 </div>
 
@@ -114,8 +91,7 @@
     <thead>
       <tr>
         <th>Member</th>
-        <th>Start Time</th>
-        <th>End Time</th>
+        <th colspan="3"></th>
       </tr>
     </thead>
     <tbody>
@@ -123,8 +99,7 @@
         <tr>
           <%if att.Hours == 0 %>
             <td><%=  Member.find(att.Member_id).First_Name%> <%= Member.find(att.Member_id).Last_Name%> </td>
-            <td><%= att.Start.strftime("%I:%M %p") %></td>
-            <td><%= att.End.strftime("%I:%M %p") %></td>
+
           <% end %>
         </tr>
       <% end %>
@@ -136,28 +111,34 @@
 <br>
 <br>
 
-<div class="d-flex align-items-center gap-3 mb-2">
-  <h1 style="margin: 0">Attendance</h1>
+<p>
+<strong>Attendance</strong>
+</p>
+<%= link_to 'New Attendance', new_attendance_path, :class => 'btn btn-primary' %>
+<table class="table">
+  <thead>
+    <tr>
+      <th>Member</th>
+      <th colspan="3"></th>
+    </tr>
+  </thead>
 
-  <%= link_to 'New Attendance', new_attendance_path, :class => 'btn btn-primary btn-new' %>
-</div>
-<div class="table-ruby table-responsive"> <%#table-responsive shows white at bottom of table %>
-  <table class="table">
-    <thead>
+  <tbody>
+    <% @EventAttendances.each do |att| %>
       <tr>
-        <th>Member</th>
-        <th>Start Time</th>
-        <th>End Time</th>
+        <%if att.Hours > 0 %>
+          <td><%=  Member.find(att.Member_id).First_Name%> <%= Member.find(att.Member_id).Last_Name%> </td>
+
+        <% end %>
       </tr>
     </thead>
+    <% end %>
 
     <tbody>
       <% @EventAttendances.each do |att| %>
         <tr>
           <%if att.Hours > 0 %>
             <td><%=  Member.find(att.Member_id).First_Name%> <%= Member.find(att.Member_id).Last_Name%> </td>
-            <td><%= att.Start.strftime("%I:%M %p") %></td>
-            <td><%= att.End.strftime("%I:%M %p") %></td>
           <% end %>
         </tr>
       <% end %>
