+++ app/views/transactions/index.html.erb
@@ -6,7 +6,7 @@
 
 <div class="d-flex align-items-center gap-3 mb-2">
   <h1 style="margin: 0">Balance </h1>
-    <h1 style="margin: 0; font-weight: 300;"><%= number_to_currency(Budget.first.Total_amount) %></h1>
+    <h1 style="margin: 0; font-weight: 300;"><%= number_to_currency(Transaction.sum(:Amount)) %></h1>
 </div>
 
 <br>
