+++ db/schema.rb
@@ -46,9 +46,7 @@ ActiveRecord::Schema.define(version: 2022_10_12_040835) do
   create_table "attendances", force: :cascade do |t|
     t.bigint "Member_id"
     t.bigint "Shift_id"
-    t.integer "Hours"
-    t.time "Start"
-    t.time "End"
+    t.integer "Hours", default: 0
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["Member_id"], name: "index_attendances_on_Member_id"
@@ -68,6 +66,7 @@ ActiveRecord::Schema.define(version: 2022_10_12_040835) do
     t.string "Name"
     t.time "Start"
     t.time "End"
+    t.string "url"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
   end
