+++ spec/models/attendance_spec.rb
@@ -9,13 +9,13 @@ RSpec.describe Attendance, type: :model do
   end
   context 'validation test' do
     it 'ensures member id' do
-      attendance = Attendance.new(Start: Time.parse("12:00"), Shift_id: 2)
+      attendance = Attendance.new(Shift_id: 2)
       expect(attendance.valid?).to eq(false)
     end
   end
   context 'validation test' do
     it 'ensures shift id prescence' do
-      attendance = Attendance.new(Member_id: 1, Start: Time.parse("12:00"))
+      attendance = Attendance.new(Member_id: 1)
       expect(attendance.valid?).to eq(false)
     end
   end
